feat: reuse tooltip code for fcs and nutrition
