+++ src/components/Map/NutritionMap/NutritionChoropleth.tsx
@@ -1,4 +1,3 @@
-import { Feature } from 'geojson';
 import L from 'leaflet';
 import { useTheme } from 'next-themes';
 import React, { useEffect, useRef, useState } from 'react';
@@ -6,7 +5,6 @@ import { GeoJSON, useMap } from 'react-leaflet';
 
 import { useSelectedCountryId } from '@/domain/contexts/SelectedCountryIdContext';
 import { CountryMapData } from '@/domain/entities/country/CountryMapData.ts';
-import { LayerWithFeature } from '@/domain/entities/map/LayerWithFeature.ts';
 import { NutrientType } from '@/domain/enums/NutrientType.ts';
 import { useNutritionQuery } from '@/domain/hooks/globalHooks';
 import NutritionChoroplethProps from '@/domain/props/NutritionChoroplethProps';
@@ -38,37 +36,7 @@ export default function NutritionChoropleth({ data, countryId, onDataUnavailable
   // the tooltip is not shown if the country is selected or there is no data available for the country
   useEffect(() => {
     if (!geoJsonRef.current || !nutritionData || !map) return () => {};
-
-    const disableTooltips = () => {
-      geoJsonRef.current?.eachLayer((layer) => {
-        if (layer instanceof L.Path) {
-          const layerElement = layer.getElement();
-          if (layerElement && !layerElement.matches(':hover')) {
-            layer.unbindTooltip();
-          }
-        }
-      });
-    };
-
-    const enableTooltips = () => {
-      geoJsonRef.current?.eachLayer((layer: LayerWithFeature) => {
-        const feature = layer.feature as Feature;
-        if (NutritionChoroplethOperations.checkIfActive(data.features[0] as CountryMapData, nutritionData)) {
-          const tooltipContainer = MapOperations.createCountryNameTooltipElement(feature?.properties?.adm0_name);
-          layer.bindTooltip(tooltipContainer, { className: 'leaflet-tooltip', sticky: true });
-        }
-      });
-    };
-
-    enableTooltips();
-
-    map.on('dragstart', disableTooltips);
-    map.on('dragend', enableTooltips);
-
-    return () => {
-      map.off('dragstart', disableTooltips);
-      map.off('dragend', enableTooltips);
-    };
+    return MapOperations.handleCountryTooltip(geoJsonRef, map, undefined, nutritionData, data);
   }, [selectedCountryId, nutritionData]);
 
   return (
